Create models folder
