+++ src/qibo/__init__.py
@@ -4,5 +4,5 @@ from qibo.backends import set_precision, set_backend, set_device
 from qibo.backends import get_backend, get_precision, get_device
 from qibo.backends import numpy_matrices as matrices
 from qibo.backends import K
-from qibo import callbacks, evolution, gates, hamiltonians, models
+from qibo import callbacks, gates, hamiltonians, models
 from qibo import parallel, optimizers, solvers
