+++ src/qibo/models/__init__.py
@@ -1,0 +1,3 @@
+from qibo.models.circuit import Circuit, QFT
+from qibo.models.evolution import StateEvolution, AdiabaticEvolution
+from qibo.models.variational import VQE, QAOA
